iso-9899: fix Windows LLP64 long import via pointer-width C shim return types
On Windows (LLP64) C `long`/`unsigned long` are 32-bit, so `iso9899_atol`,
`iso9899_strtol`, and `iso9899_strtoul` imported as returning Int32/UInt32 — not
matching the Swift wrappers' `Int`/`UInt` returns (`cannot convert Int32 to Int`).
On LP64 (macOS/Linux) C `long` is 64-bit = `Int`, so it compiled there and the bug
was latent.

Fix by changing the C shim return types to pointer-width intptr_t/uintptr_t (rather
than adding Int()/UInt() conversions in Swift). intptr_t/uintptr_t are pointer-width
on every platform, so they import as Swift `Int`/`UInt` uniformly — like ssize_t,
which never had this problem. The Swift wrappers keep their natural `Int`/`UInt`
return types with zero conversions and no public-API change; the value-preserving
widening of the platform `long` result happens in C.

Verified: swift build clean on macOS (no regression). Unblocks the Windows build of
the swift-parser-primitives / swift-file-system dependency graph.

+++ Sources/CISO9899Stdlib/include/iso9899_stdlib.h
@@ -9,12 +9,21 @@
 #define ISO9899_STDLIB_H
 
 #include <stdlib.h>
+#include <stdint.h>
 
 // String conversion functions (Section 7.22.1)
-// These are typically real functions, but we wrap for consistency
+// These are typically real functions, but we wrap for consistency.
+//
+// `long`-returning conversions (atol / strtol / strtoul) return via pointer-width
+// intptr_t/uintptr_t rather than C `long`. C `long` is 64-bit on LP64 (macOS/Linux)
+// but 32-bit on LLP64 (Windows), so it imports into Swift as `Int`/`UInt` on LP64
+// but `Int32`/`UInt32` on Windows — mismatching the Swift wrappers' `Int`/`UInt`.
+// intptr_t/uintptr_t are pointer-width on every platform, so they import as Swift
+// `Int`/`UInt` uniformly (like ssize_t), letting the wrappers stay conversion-free.
+// The C-level widening of the platform `long` result is value-preserving.
 static inline double iso9899_atof(const char *nptr) { return atof(nptr); }
 static inline int iso9899_atoi(const char *nptr) { return atoi(nptr); }
-static inline long iso9899_atol(const char *nptr) { return atol(nptr); }
+static inline intptr_t iso9899_atol(const char *nptr) { return atol(nptr); }
 static inline long long iso9899_atoll(const char *nptr) { return atoll(nptr); }
 
 static inline double iso9899_strtod(const char *nptr, char **endptr) {
@@ -26,13 +35,13 @@ static inline float iso9899_strtof(const char *nptr, char **endptr) {
 static inline long double iso9899_strtold(const char *nptr, char **endptr) {
     return strtold(nptr, endptr);
 }
-static inline long iso9899_strtol(const char *nptr, char **endptr, int base) {
+static inline intptr_t iso9899_strtol(const char *nptr, char **endptr, int base) {
     return strtol(nptr, endptr, base);
 }
 static inline long long iso9899_strtoll(const char *nptr, char **endptr, int base) {
     return strtoll(nptr, endptr, base);
 }
-static inline unsigned long iso9899_strtoul(const char *nptr, char **endptr, int base) {
+static inline uintptr_t iso9899_strtoul(const char *nptr, char **endptr, int base) {
     return strtoul(nptr, endptr, base);
 }
 static inline unsigned long long iso9899_strtoull(const char *nptr, char **endptr, int base) {
